+++ packages/collector/test/nativeModuleRetry/test.js
@@ -6,14 +6,11 @@
 'use strict';
 
 const async = require('async');
-const copy = require('recursive-copy');
 const { expect } = require('chai');
 const fs = require('fs');
 const os = require('os');
 const path = require('path');
 const rimraf = require('rimraf');
-const semver = require('semver');
-const tar = require('tar');
 
 const config = require('../../../core/test/config');
 const { retry } = require('../../../core/test/test_util');
@@ -21,7 +18,6 @@ const ProcessControls = require('../test_util/ProcessControls');
 const globalAgent = require('../globalAgent');
 
 const sharedMetricsNodeModules = path.join(__dirname, '..', '..', '..', 'shared-metrics', 'node_modules');
-const resourcesPath = path.join(__dirname, 'resources');
 
 describe('retry loading native addons', function () {
   const timeout = Math.max(config.getTestTimeout(), 20000);
@@ -37,9 +33,6 @@ describe('retry loading native addons', function () {
       nodeModulesPath: sharedMetricsNodeModules,
       nativeModulePath: path.join(sharedMetricsNodeModules, 'event-loop-stats'),
       backupPath: path.join(os.tmpdir(), 'event-loop-stats-backup'),
-      resourcesPath,
-      corruptTarGzPath: path.join(resourcesPath, 'event-loop-stats-corrupt.tar.gz'),
-      corruptUnpackedPath: path.join(resourcesPath, 'event-loop-stats'),
       check: ([allMetrics, aggregated]) => {
         // check that libuv stats are initially reported as unsupported
         let foundAtLeastOneUnsupported;
@@ -69,9 +62,6 @@ describe('retry loading native addons', function () {
       nodeModulesPath: sharedMetricsNodeModules,
       nativeModulePath: path.join(sharedMetricsNodeModules, 'gcstats.js'),
       backupPath: path.join(os.tmpdir(), 'gcstats.js-backup'),
-      resourcesPath,
-      corruptTarGzPath: path.join(resourcesPath, 'gcstats.js-corrupt.tar.gz'),
-      corruptUnpackedPath: path.join(resourcesPath, 'gcstats.js'),
       check: ([allMetrics, aggregated]) => {
         // check that gc stats are initially reported as unsupported
         let foundAtLeastOneUnsupported;
@@ -108,22 +98,6 @@ describe('retry loading native addons', function () {
       runCopyPrecompiledForNativeAddonTest.bind(this, agentControls, controls, retryTimeout)
     );
   });
-
-  const mochaSuiteFn = semver.gte(process.versions.node, '10.0.0') ? describe : describe.skip;
-  mochaSuiteFn('metrics are activated lazily by compiling on demand when they are initially missing', () => {
-    const controls = new ProcessControls({
-      appPath: path.join(__dirname, 'app'),
-      useGlobalAgent: true,
-      env: {
-        INSTANA_COPY_PRECOMPILED_NATIVE_ADDONS: 'false',
-        INSTANA_REBUILD_NATIVE_ADDONS_ON_DEMAND: 'true'
-      }
-    }).registerTestHooks();
-
-    metricAddonsTestConfigs.forEach(
-      runCompileOnDemandForNativeAddonTest.bind(this, agentControls, controls, retryTimeout)
-    );
-  });
 });
 
 function runCopyPrecompiledForNativeAddonTest(agentControls, controls, retryTimeout, opts) {
@@ -139,47 +113,6 @@ function runCopyPrecompiledForNativeAddonTest(agentControls, controls, retryTime
         [
           //
           rimraf.bind(null, opts.nativeModulePath),
-          fs.rename.bind(null, opts.backupPath, opts.nativeModulePath)
-        ],
-        done
-      );
-    });
-
-    it('metrics from native add-ons should become available at some point', () =>
-      retry(
-        () =>
-          Promise.all([
-            //
-            agentControls.getAllMetrics(controls.getPid()),
-            agentControls.getAggregatedMetrics(controls.getPid()),
-            agentControls.getEvents()
-          ]).then(opts.check),
-        retryTimeout
-      ));
-  });
-}
-
-function runCompileOnDemandForNativeAddonTest(agentControls, controls, retryTimeout, opts) {
-  describe(opts.name, () => {
-    before(done => {
-      async.series(
-        [
-          tar.x.bind(null, {
-            cwd: opts.resourcesPath,
-            file: opts.corruptTarGzPath
-          }),
-          fs.rename.bind(null, opts.nativeModulePath, opts.backupPath),
-          copy.bind(null, opts.corruptUnpackedPath, opts.nativeModulePath)
-        ],
-        done
-      );
-    });
-
-    after(done => {
-      async.series(
-        [
-          rimraf.bind(null, opts.nativeModulePath),
-          rimraf.bind(null, opts.corruptUnpackedPath),
           fs.rename.bind(null, opts.backupPath, opts.nativeModulePath)
         ],
         done