+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - An issue event is sent if the application uses an EOL (end of life) version of Node.js. Applicable only for non serverless environments
 - Manual spans creation handles tags that are passed as non extensible objetcts - when `Object.freeze` or `Object.preventExtensions` is applied to the tags
+- [AWS SDKv2] Instrumentation of SNS
 
 ## 1.135.0